completion of exercises 4.15 to 4.23 for bloglist app

+++ part4/bloglist/controllers/blogs.js
@@ -1,16 +1,15 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
+const { userExtractor } = require('../utils/middleware');
   
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   response.json(blogs);
 });
-  
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body;
 
-  const user = await User.findById(body.userId);
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body;
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -27,7 +26,14 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  const userId = request.user?._id;
+
+  if (userId.toString() !== blog?.user?.toString()) {
+    return response.status(401).json({ error: 'you must be the author of the blog to delete it' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
